Move encryption to its own package & add new encryption functions
