+++ src/main/resources/reference.conf
@@ -34,10 +34,10 @@ write-format = ${?COMET_WRITE_FORMAT}
 merge-force-distinct = false
 merge-force-distinct = ${?COMET_MERGE_FORCE_DISTINCT}
 
-file-system = "gs://nps-datastore"
-#file-system = "hdfs://localhost:9000"
+#file-system = "gs://nps-datastore"
+file-system = "hdfs://localhost:9000"
 #file-system = "abfs://cometfs@ebizcomet.dfs.core.windows.net"
-file-system = ${?COMET_FS}
+#file-system = ${?COMET_FS}
 
 chewer-prefix = "comet.chewer"
 chewer-prefix = ${?COMET_CHEWER_PREFIX}
@@ -266,12 +266,12 @@ area {
 
 privacy {
   options = {
-    "none": "com.ebiznext.comet.utils.No",
-    "hide": "com.ebiznext.comet.utils.Hide",
-    "md5": "com.ebiznext.comet.utils.Md5",
-    "sha1": "com.ebiznext.comet.utils.Sha1",
-    "sha256": "com.ebiznext.comet.utils.Sha256",
-    "sha512": "com.ebiznext.comet.utils.Sha512"
+    "none": "com.ebiznext.comet.privacy.No",
+    "hide": "com.ebiznext.comet.privacy.Hide",
+    "md5": "com.ebiznext.comet.privacy.Md5",
+    "sha1": "com.ebiznext.comet.privacy.Sha1",
+    "sha256": "com.ebiznext.comet.privacy.Sha256",
+    "sha512": "com.ebiznext.comet.privacy.Sha512"
   }
 }
 
@@ -317,7 +317,7 @@ spark {
   #  yarn.principal = ${?SPARK_YARN_PRINCIPAL}
   #  yarn.keytab = ${?SPARK_YARN_KEYTAB}
   debug.maxToStringFields = 100
-  #master = "local[*]"
+  master = "local[*]"
   #  sql.catalogImplementation="hive"
   #  sql.catalogImplementation="in-memory"
 }
