+++ src/main/scala/com/ebiznext/comet/privacy/Encryption.scala
@@ -1,4 +1,4 @@
-package com.ebiznext.comet.utils
+package com.ebiznext.comet.privacy
 
 /**
   * Several encryption methods used in privacy management
@@ -45,3 +45,17 @@ object Hide extends Encryption {
 object No extends Encryption {
   def encrypt(s: String): String = s
 }
+
+object Initials extends Encryption {
+  def encrypt(s: String): String = {
+    s.split("\\s+").map(_.substring(0,1)).mkString(".")
+  }
+}
+
+object EmailMD5 extends Encryption {
+  def encrypt(s: String): String = {
+    val split = s.split('@')
+    Md5.encrypt(split(0)) + "@" + split(1)
+  }
+}
+
