+++ src/main/scala/com/ebiznext/comet/schema/model/PrivacyLevel.scala
@@ -23,7 +23,7 @@ package com.ebiznext.comet.schema.model
 import java.util.Locale
 
 import com.ebiznext.comet.config.Settings
-import com.ebiznext.comet.utils.Encryption
+import com.ebiznext.comet.privacy.Encryption
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.annotation.{JsonDeserialize, JsonSerialize}
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer
